Divided Settings dialog into files
