+++ shortcut_composer/composer_utils/settings_dialog_utils/__init__.py
@@ -1,0 +1,4 @@
+from .comboboxes import Comboboxes
+from .forms import Forms
+
+__all__ = ["Comboboxes", "Forms"]
